actor system inspection and separate module
